move implementation to curve analysis
