+++ releasenotes/notes/curve-analysis-02a702a81e014adf.yaml
@@ -78,6 +78,13 @@ upgrade:
     more than three curves. The corresponding fit model is dynamically generated
     based on new fit option ``reps`` in the associated analysis :class:`DragCalAnalysis`.
     This may sometimes provide better accuracy for estimating the DRAG ``beta`` parameter.
+  - |
+    Implementation of several methods in :class:`.BaseCurveAnalysis` has been moved to
+    its subclass :class:`.CurveAnalysis`. :meth:`.BaseCurveAnalysis._run_data_processing`,
+    :meth:`.BaseCurveAnalysis._format_data`, and :meth:`.BaseCurveAnalysis._run_curve_fit`,
+    have been turned into abstract method, and :meth:`.BaseCurveAnalysis._generate_fit_guesses`
+    has been moved to :class:`CurveAnalysis`. There is no net upgrade on the behavior of
+    curve analysis subclasses, since :class:`.BaseCurveAnalysis` is an abstract class.
 deprecations:
   - |
     Multiple methods, classes and functions in curve analysis have been deprecated and
